# Conflicts:
#	src/main/resources/application.yml

+++ src/main/java/edu/ntnu/idatt2106_09/backend/model/GroceryItemRecipe.java
@@ -27,7 +27,7 @@ public class GroceryItemRecipe {
     private GroceryItem groceryItem;
 
     @Column(name = "amount")
-    private int amount;
+    private double amount;
 
     public GroceryItemRecipe(Recipe recipe, GroceryItem groceryItem) {
         this.recipe = recipe;
